+++ lib/Screens/splash.dart
@@ -1,0 +1,40 @@
+import 'package:flutter/material.dart';
+import 'dart:async';
+import 'second_screen.dart';
+
+class Splashscreen extends StatefulWidget {
+  @override
+  _SplashscreenState createState() => _SplashscreenState();
+}
+
+class _SplashscreenState extends State<Splashscreen> {
+  @override
+  void initState() {
+    super.initState();
+    startTime();
+  }
+
+  startTime() async {
+    var duration = Duration(seconds: 6);
+    return Timer(duration, navigateToDeviceScreen);
+  }
+
+  navigateToDeviceScreen() {
+    Navigator.pushReplacement(
+        context, MaterialPageRoute(builder: (context) => const MyHomePage()));
+  }
+
+  @override
+  Widget build(BuildContext context){
+    return Scaffold(
+      body: Center(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            SizedBox(
+              width: 300,
+              child: Image.asset('lib/assets/Mortar.png', width: 430,),)
+          ],)),
+    );
+  }
+}
